Fixed protos.

+++ mess/includes/a7800.h
@@ -17,7 +17,7 @@ extern unsigned char *a7800_ram;
 extern unsigned char *a7800_cartridge_rom;
 MACHINE_INIT( a7800 );
 MACHINE_INIT( a7800p );
-UINT32 a7800_partialcrc(const unsigned char *,unsigned int);
+UINT32 a7800_partialcrc(const unsigned char *, size_t);
 int a7800_cart_load(int id, mame_file *cartfile, int open_mode);
 int a7800p_cart_load(int id, mame_file *cartfile, int open_mode);
 READ_HANDLER  ( a7800_TIA_r );
